Not sure if I like the new scene hiararchy but there are matrixes now and also preformance is wahoo!

+++ EpsilonEngine/EpsilonEngine/Data Types/KeyboardButton.cs
@@ -1,6 +1,6 @@
 ﻿namespace EpsilonEngine
 {
-    public enum KeyboardButton : int
+    public enum KeyboardButton : byte
     {
         A,
         B,
